+++ src/lib/ledger.h
@@ -26,7 +26,7 @@ ledger_status ledger_open_context(ledger_ctx *ctx, const char *root_directory);
 ledger_status ledger_open_topic(ledger_ctx *ctx,
                                 const char *name,
                                 const unsigned int *partition_ids,
-                                unsigned int partition_count,
+                                size_t partition_count,
                                 ledger_topic_options *options);
 ledger_status ledger_write(ledger_ctx *ctx, const char *topic_name,
                            const char *partition_key, size_t key_len,
