+++ src/lib/topic.c
@@ -42,7 +42,7 @@ ledger_status ledger_topic_options_init(ledger_topic_options *options) {
 
 ledger_status ledger_topic_open(ledger_topic *topic, const char *root,
                                 const unsigned int *partition_ids,
-                                unsigned int partition_count,
+                                size_t partition_count,
                                 ledger_topic_options *options) {
     int i;
     ledger_status rc;
